Move st.text out of markdown
