+++ lib/streamlit/elements/markdown.py
@@ -1,5 +1,4 @@
 from streamlit.proto import Markdown_pb2
-from streamlit.proto import Text_pb2
 from streamlit import type_util
 from .utils import _clean_text
 
@@ -208,28 +207,4 @@ class LatexMixin:
         latex_proto = Markdown_pb2.Markdown()
         latex_proto.body = "$$\n%s\n$$" % _clean_text(body)
         dg._enqueue("markdown", latex_proto)
-
-
-# TODO: This doesn't belong here
-class TextMixin:
-    def text(dg, body):
-        """Write fixed-width and preformatted text.
-
-        Parameters
-        ----------
-        body : str
-            The string to display.
-
-        Example
-        -------
-        >>> st.text('This is some text.')
-
-        .. output::
-           https://share.streamlit.io/0.25.0-2JkNY/index.html?id=PYxU1kee5ubuhGR11NsnT1
-           height: 50px
-
-        """
-        text_proto = Text_pb2.Text()
-        text_proto.body = _clean_text(body)
-        dg._enqueue("text", text_proto)
